# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API Server for my Application",
    "title": "Moscow Sights API",
    "contact": {},
    "version": "1.0"
  },
  "host": "localhost:8000",
  "basePath": "/",
  "paths": {
    "/api/lists": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "get all lists",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["lists"],
        "summary": "Get All Lists",
        "operationId": "get-all-lists",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/handler.allListsResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/handler.errorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/handler.errorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/handler.errorResponse"
            }
          },
          "default": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/handler.errorResponse"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "create sights list",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["lists"],
        "summary": "Create sights list",
        "operationId": "create-list",
        "parameters": [
          {
            "description": "list info",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/sights.TodoList"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "integer"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/handler.errorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/handler.errorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/handler.errorResponse"
            }
          },
          "default": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/handler.errorResponse"
            }
          }
        }
      }
    },
    "/api/lists/:id": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "get list by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["lists"],
        "summary": "Get List By Id",
        "operationId": "get-list-by-id",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/sights.ListItem"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/handler.errorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/handler.errorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/handler.errorResponse"
            }
          },
          "default": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/handler.errorResponse"
            }
          }
        }
      }
    },
    "/auth/sign-in": {
      "post": {
        "description": "login",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["auth"],
        "summary": "SignIn",
        "operationId": "login",
        "parameters": [
          {
            "description": "credentials",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/handler.SignInRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "token",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/handler.errorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/handler.errorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/handler.errorResponse"
            }
          },
          "default": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/handler.errorResponse"
            }
          }
        }
      }
    },
    "/auth/sign-up": {
      "post": {
        "description": "create account",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["auth"],
        "summary": "SignUp",
        "operationId": "create-account",
        "parameters": [
          {
            "description": "account info",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/sights.User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "integer"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/handler.errorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/handler.errorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/handler.errorResponse"
            }
          },
          "default": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/handler.errorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "handler.SignInRequest": {
      "type": "object",
      "required": ["password", "username"],
      "properties": {
        "password": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
    "handler.allListsResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/sights.TodoList"
          }
        }
      }
    },
    "handler.errorResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "sights.ListItem": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "item_id": {
          "type": "string"
        },
        "list_id": {
          "type": "string"
        }
      }
    },
    "sights.TodoList": {
      "type": "object",
      "required": ["title"],
      "properties": {
        "description": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        }
      }
    },
    "sights.User": {
      "type": "object",
      "required": ["name", "password", "username"],
      "properties": {
        "name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}